chore: Update batch size and num_workers in dataset configuration

+++ tools/train_val.py
@@ -21,7 +21,7 @@ from lib.helpers.scheduler_helper import build_lr_scheduler
 from lib.helpers.trainer_helper import Trainer
 from lib.helpers.tester_helper import Tester
 from lib.helpers.utils_helper import create_logger
-#from lib.helpers.utils_helper import set_random_seed
+from lib.helpers.utils_helper import set_random_seed
 from accelerate.utils import set_seed
 from accelerate import Accelerator, DistributedDataParallelKwargs
 
@@ -32,16 +32,20 @@ args = parser.parse_args()
 
 
 def main():
+    accelerator = Accelerator(kwargs_handlers=[DistributedDataParallelKwargs(find_unused_parameters=True)])
     assert (os.path.exists(args.config))
     cfg = yaml.load(open(args.config, 'r'), Loader=yaml.Loader)
-    set_seed(cfg.get('random_seed', 444))
+    torch.backends.cudnn.benchmark = False
+    torch.backends.cudnn.deterministic = True
+    set_seed(cfg.get('random_seed', 444),device_specific=True, deterministic=False)
 
     model_name = cfg['model_name']
-    timestamp = datetime.datetime.now().strftime('%Y%m%d_%H%M%S')
+    timestamp = datetime.datetime.now().strftime('%Y%m%d_%H%M')
     output_path = os.path.join('./' + cfg["trainer"]['save_path'], f"{model_name}_{timestamp}")
-    os.makedirs(output_path, exist_ok=True)
+    if not os.path.exists(output_path):
+        os.makedirs(output_path, exist_ok=True)
 
-    shutil.copy(args.config, output_path)
+        shutil.copy(args.config, output_path)
 
     log_file = os.path.join(output_path, f'train.log.{timestamp}')
     logger = create_logger(log_file)
@@ -63,7 +67,6 @@ def main():
     # else:
     #     model = torch.nn.DataParallel(model, device_ids=gpu_ids).to(device)
     #     loss = torch.nn.DataParallel(loss, device_ids=gpu_ids).to(device)
-    accelerator = Accelerator(kwargs_handlers=[DistributedDataParallelKwargs(find_unused_parameters=True)])
     # device = accelerator.device
     #model = model.to(device)
